+++ src/main/java/gov/nasa/pds/registry/mgr/cmd/dd/LoadDDCmd.java
@@ -5,14 +5,15 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import gov.nasa.pds.registry.common.dd.LddUtils;
+import gov.nasa.pds.registry.common.es.dao.DataLoader;
+import gov.nasa.pds.registry.common.es.dao.dd.DataDictionaryDao;
+import gov.nasa.pds.registry.common.es.service.JsonLddLoader;
 import gov.nasa.pds.registry.mgr.Constants;
 import gov.nasa.pds.registry.mgr.cfg.RegistryCfg;
 import gov.nasa.pds.registry.mgr.cmd.CliCommand;
-import gov.nasa.pds.registry.mgr.dao.DataLoader;
 import gov.nasa.pds.registry.mgr.dao.RegistryManager;
 import gov.nasa.pds.registry.mgr.dd.CsvLddLoader;
-import gov.nasa.pds.registry.mgr.dd.JsonLddLoader;
-import gov.nasa.pds.registry.mgr.dd.LddUtils;
 
 
 /**
@@ -129,7 +130,8 @@ public class LoadDDCmd implements CliCommand
         }
 
         // Init LDD loader
-        JsonLddLoader loader = new JsonLddLoader(cfg.url, cfg.indexName, cfg.authFile);
+        DataDictionaryDao ddDao = RegistryManager.getInstance().getDataDictionaryDao();
+        JsonLddLoader loader = new JsonLddLoader(ddDao, cfg.url, cfg.indexName, cfg.authFile);
         loader.loadPds2EsDataTypeMap(LddUtils.getPds2EsDataTypeCfgFile());
 
         //Load LDD
