+++ src/test/java/tt/TestDataDictionaryDao.java
@@ -3,11 +3,11 @@ package tt;
 import java.util.Collections;
 import java.util.List;
 
+import gov.nasa.pds.registry.common.es.dao.dd.DataDictionaryDao;
+import gov.nasa.pds.registry.common.es.dao.dd.LddInfo;
+import gov.nasa.pds.registry.common.es.dao.dd.LddVersions;
 import gov.nasa.pds.registry.mgr.cfg.RegistryCfg;
 import gov.nasa.pds.registry.mgr.dao.RegistryManager;
-import gov.nasa.pds.registry.mgr.dao.dd.DataDictionaryDao;
-import gov.nasa.pds.registry.mgr.dao.dd.LddInfo;
-import gov.nasa.pds.registry.mgr.dao.dd.LddVersions;
 
 
 public class TestDataDictionaryDao
