+++ src/test/java/tt/TestSchemaUpdater.java
@@ -4,7 +4,6 @@ import java.io.File;
 import gov.nasa.pds.registry.mgr.cfg.RegistryCfg;
 import gov.nasa.pds.registry.mgr.dao.RegistryManager;
 import gov.nasa.pds.registry.mgr.dao.schema.SchemaUpdater;
-import gov.nasa.pds.registry.mgr.dd.JsonLddLoader;
 
 
 public class TestSchemaUpdater
@@ -18,14 +17,13 @@ public class TestSchemaUpdater
     
     public static void testUpdateLdds() throws Exception
     {
-        JsonLddLoader lddLoader = new JsonLddLoader("http://localhst:9200", "registry", null);
-        lddLoader.loadPds2EsDataTypeMap(new File("src/main/resources/elastic/data-dic-types.cfg"));
-
+        
         RegistryCfg cfg = new RegistryCfg();
         cfg.url = "http://localhst:9200";
         cfg.indexName = "registry";
         
         RegistryManager.init(cfg);
+        
         try
         {
             SchemaUpdater updater = new SchemaUpdater(cfg, false);
